requirements.txt
pydantic
fastapi
pyjwt
pymongo
uvicorn
bcrypt
